enabling reset button for collection settings

+++ core/static/js/collectionInfo.js
@@ -14,12 +14,26 @@ function setupVideoSearch() {
   });
 }
 
+function getCollectionIdValue() {
+  const collectionForm = document.getElementById("collection-settings-form");
+  const idInput = collectionForm.querySelector("input[name='id']");
+  const idValue = idInput?.value;
+  if (idValue === undefined) {
+    console.error("Failed to get collection id from form");
+    return;
+  }
+  return idValue;
+}
+
 function setupDeleteCollection() {
   const deleteButton = document.getElementById("collection-confirm-delete");
-  const collectionForm = document.getElementById("collection-settings-form");
+  const idValue = getCollectionIdValue();
   deleteButton.addEventListener("click", async () => {
-    const idInput = collectionForm.querySelector("input[name='id']");
-    const idValue = idInput.value;
+    if (idValue === undefined) {
+      deleteButton.disabled = true;
+      deleteButton.classList.add("disabled");
+      return;
+    }
     const deleteResponse = await fetch(`/collections/delete/${idValue}/`, {
       method: "DELETE",
       headers: {
@@ -34,9 +48,36 @@ function setupDeleteCollection() {
   });
 }
 
+// function setupUpdateCollectionSettings() {
+
+// }
+
+function setupResetCollectionSettings() {
+  const resetButton = document.getElementById("collection-settings-reset");
+
+  resetButton.addEventListener("click", async (event) => {
+    event.preventDefault();
+    const collectionId = getCollectionIdValue();
+    const resetResponse = await fetch(`/display-collection-settings/${collectionId}/`);
+    if (!resetResponse.ok) {
+      console.error("Failed to reset collection settings");
+      return;
+    }
+    const settingsHTML = await resetResponse.text();
+    const currentSettingsEl = document.getElementById("collection-settings");
+    currentSettingsEl.outerHTML = settingsHTML;
+    setupCollectionSettings();
+  });
+}
+
+function setupCollectionSettings() {
+  setupDeleteCollection();
+  setupResetCollectionSettings();
+}
+
 function initialize() {
   setupVideoSearch();
-  setupDeleteCollection();
+  setupCollectionSettings();
 }
 
 initialize();
